# src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: API Rest - Backend - Arena Mafia
  version: v1
  description: |
    Backend do projeto para a Arena Mafia, desenvolvido como parte do curso de
    Análise e Desenvolvimento de Sistemas do Instituto Federal de Santa Catarina.
  termsOfService: https://github.com/LucasRibasCardoso/backend-arena-mafia
  license:
    name: Apache 2.0
    url: https://github.com/LucasRibasCardoso
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Token de autenticação JWT obtido no login."
  schemas:
    FieldErrorResponseDto:
      type: object
      properties:
        fieldName: {
          type: string,
          description: "Nome do campo que falhou na validação.",
          example: "username"
        }
        message: {
          type: string,
          description: "Mensagem de erro para o campo.",
          example: "O usuário é obrigatório"
        }
      required: [ fieldName, message ]
    ErrorResponseDto:
      type: object
      properties:
        timestamp: {
          type: string,
          format: date-time,
          example: "2025-08-15T13:55:00Z"
        }
        status: {
          type: integer,
          example: 400
        }
        message: {
          type: string,
          example: "Requisição inválida."
        }
        path: {
          type: string,
          example: "/api/some/endpoint"
        }
        fieldErrors:
          type: array
          description: "Lista de erros de validação de campos, se houver."
          items: { $ref: "#/components/schemas/FieldErrorResponseDto" }
          nullable: true
      required: [ timestamp, status, message, path, fieldErrors ]
...
    ValidateOtpRequestDto:
      type: object
      description: "Dados para verificação da conta via código OTP."
      required: [ phone, code ]
      properties:
        phone:
          type: string
          description: "Telefone de usuário da conta a ser verificada."
          example: "+5547912345678"
        code:
          type: string
          description: "Código OTP de 6 dígitos recebido por SMS."
          example: "123456"
...
    SimpleMessageResponseDto:
      type: object
      properties:
        message:
          type: string
          example: "Operação realizada com sucesso."
    ForgotPasswordRequestDto:
      type: object
      description: "Telefone do usuário que está solicitando a alteração da senha."
      required: [ phone ]
      properties:
        phone:
          type: string
          description: "Telefone de usuário da conta para a qual a senha deve ser alterada."
          example: "+5547912345678"
    PasswordResetTokenResponseDto:
      type: object
      properties:
        passwordResetToken:
          type: string
          description: "Token temporário para redefinição de senha."
          example: "dawda-1234-asdas-5678-asdasd-9876"
    ResetPasswordRequestDto:
      type: object
      description: "Dados necessários para a etapa final da redefinição de senha."
      required: [ passwordResetToken, newPassword, confirmPassword ]
      properties:
        passwordResetToken:
          type: string
          description: "O token seguro de uso único recebido na etapa de validação do OTP."
          example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
        newPassword:
          type: string
          format: "password"
          description: "A nova senha do usuário. Deve ter entre 6 e 20 caracteres e não pode conter espaços."
          example: "novaSenhaSegura123"
        confirmPassword:
          type: string
          format: "password"
          description: "Confirmação da nova senha. Deve ser idêntica ao campo 'newPassword'."
          example: "novaSenhaSegura123"
...
  responses:
    ErrorResponseInternalServerError:
      description: Erro interno do servidor.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponseDto"
          example:
            timestamp: "2025-08-15T13:55:00Z"
            status: 500
            message: "Erro interno do servidor. Por favor, tente novamente mais tarde."
            path: "/api/some/endpoint"
...
paths:
...
  /api/auth/forgot-password:
    post:
      operationId: "forgotPassword"
      tags:
        - "Autenticação"
      summary: "Iniciar redefinição de senha"
      description: "Inicia o fluxo de redefinição de senha. Se o usuário existir e estiver ativo, um código OTP é enviado para o seu telefone. Por segurança, a resposta é sempre a mesma para evitar a enumeração de contas."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequestDto"
      responses:
        "202":
          description: "Solicitação processada. A mensagem não indica se o telefone foi encontrado, para proteger a privacidade do usuário."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleMessageResponseDto"
              example:
                message: "Se o telefone informado for válido, enviaremos um código para verificação da conta."
        "400":
          description: "Requisição inválida. Ocorre se o formato do telefone for inválido."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              examples:
                MissingPhone:
                  summary: "Número de telefone não informado"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 400
                    message: "Erro de validação. Verifique os campos informados."
                    path: "/api/auth/forgot-password"
                    fieldErrors:
                      - fieldName: "phone"
                        message: "Por favor, informe seu telefone."
                InvalidPhone:
                  summary: "Número de telefone inválido"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 400
                    message: "Número de telefone inválido. Verifique o DDD e a quantidade de dígitos."
                    path: "/api/auth/forgot-password"
        "409":
          description: "Conflito de estado. Ocorre se a conta não estiver ativada"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              examples:
                PendingVerification:
                  summary: "Conta pendente de verificação"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 409
                    message: "Atenção: Você precisa ativar sua conta. Por favor, termine o processo de cadastro."
                    path: "/api/auth/forgot-password"
                Locked:
                  summary: "Conta bloqueada"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 409
                    message: "Atenção: Sua conta está bloqueada. Por favor, contate o suporte."
                    path: "/api/auth/forgot-password"
                Disabled:
                  summary: "Conta desativada"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 409
                    message: "Atenção: Sua conta está desativada."
                    path: "/api/auth/forgot-password"
        "500":
          $ref: "#/components/responses/ErrorResponseInternalServerError"
  /api/auth/password-reset-token:
    post:
      operationId: "passwordResetToken"
      tags:
        - "Autenticação"
      summary: "Validar código e gerar token de redefinição de senha"
      description: "Segunda etapa do fluxo de redefinição de senha. Valida o código OTP recebido pelo usuário. Se o código for válido, gera e retorna um token temporário de uso único para a redefinição de senha."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateOtpRequestDto"
      responses:
        "200":
          description: "Código OTP validado com sucesso. Retorna o token temporário para redefinição de senha."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetTokenResponseDto"
        "400":
          description: "Requisição inválida. Ocorre se houver erros de validação (código em branco) ou se o código OTP for inválido/expirado."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              examples:
                MissingCode:
                  summary: "Código de verificação ausente"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 400
                    message: "Erro de validação. Verifique os campos informados."
                    path: "/api/auth/password-reset-token"
                    fieldErrors:
                      - fieldName: "code"
                        message: "O código de verificação é obrigatório."
                InvalidPhone:
                  summary: "Número de telefone inválido"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 400
                    message: "Número de telefone inválido. Verifique o DDD e a quantidade de dígitos."
                    path: "/api/auth/password-reset-token"
                InvalidCode:
                  summary: "Código OTP inválido ou expirado"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 400
                    message: "Código de verificação inválido ou expirado."
                    path: "/api/auth/password-reset-token"
        "404":
          description: "Não encontrado. Ocorre se não houver um usuário associado ao telefone fornecido."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              example:
                timestamp: "2025-08-15T13:55:00Z"
                status: 404
                message: "Usuário não encontrado. Verifique o número de telefone informado e tente novamente."
                path: "/api/auth/password-reset-token"
        "409":
          description: "Conflito de estado. Ocorre se a conta não estiver ativada"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
            examples:
              PendingVerification:
                summary: "Conta pendente de verificação"
                value:
                  timestamp: "2025-08-15T13:55:00Z"
                  status: 409
                  message: "Atenção: Você precisa ativar sua conta. Por favor, termine o processo de cadastro."
                  path: "/api/auth/password-reset-token"
              Locked:
                summary: "Conta pendente de verificação"
                value:
                  timestamp: "2025-08-15T13:55:00Z"
                  status: 409
                  message: "Atenção: Sua conta está bloqueada. Por favor, contate o suporte."
                  path: "/api/auth/password-reset-token"
              Disabled:
                summary: "Conta pendente de verificação"
                value:
                  timestamp: "2025-08-15T13:55:00Z"
                  status: 409
                  message: "Atenção: Sua conta está desativada."
                  path: "/api/auth/password-reset-token"
        "500":
          $ref: "#/components/responses/ErrorResponseInternalServerError"
  /api/auth/reset-password:
    post:
      operationId: "resetPassword"
      tags:
        - "Autenticação"
      summary: "Redefinir a senha do usuário"
      description: "Etapa final do fluxo de redefinição de senha. Recebe o token temporário gerado na etapa anterior e a nova senha para efetivar a alteração no sistema."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequestDto"
      responses:
        "200":
          description: "Senha redefinida com sucesso."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleMessageResponseDto"
              example:
                message: "Senha redefinida com sucesso. Você já pode fazer o login com sua nova senha."
        "400":
          description: "Requisição inválida. Ocorre se o token de redefinição for inválido/expirado, as senhas não coincidirem ou houver erros de validação."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              examples:
                MissingToken:
                  summary: "Token de redefinição ausente"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 400
                    message: "Erro de validação. Verifique os campos informados."
                    path: "/api/auth/reset-password"
                    fieldErros:
                      - fieldName: "passwordResetToken"
                        message: "O token de redefinição de senha é obrigatório."
                PasswordsNotMatch:
                  summary: "Senhas não coincidem"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 400
                    message: "Erro de validação. Verifique os campos informados."
                    path: "/api/auth/reset-password"
                    fieldErros:
                      - fieldName: "confirmedPassword"
                        message: "A senha de confirmação não corresponde à nova senha."

                InvalidToken:
                  summary: "Token de redefinição inválido ou expirado"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 400
                    message: "Token inválido ou expirado."
                    path: "/api/auth/reset-password"
        "404":
          description: "Não encontrado. Ocorre se o token de redefinição não estiver associado a nenhum usuário válido."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              example:
                timestamp: "2025-08-15T13:55:00Z"
                status: 400
                message: "Ocorreu um erro ao redefinir sua senha. Por favor, inicie o processo novamente."
                path: "/api/auth/reset-password"
        "409":
          description: "Conflito de estado. Ocorre se a conta não estiver ativada"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              examples:
                PendingVerification:
                  summary: "Conta pendente de verificação"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 409
                    message: "Atenção: Você precisa ativar sua conta. Por favor, termine o processo de cadastro."
                    path: "/api/auth/reset-password"
                Locked:
                  summary: "Conta pendente de verificação"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 409
                    message: "Atenção: Sua conta está bloqueada. Por favor, contate o suporte."
                    path: "/api/auth/reset-password"
                Disabled:
                  summary: "Conta pendente de verificação"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 409
                    message: "Atenção: Sua conta está desativada."
                    path: "/api/auth/reset-password"
        "500":
          $ref: "#/components/responses/ErrorResponseInternalServerError"

  # ---- Endpoints atualização da conta e perfil ----
...